플레이어 스킬Json으로 추가, 개발자용 어둠의경로 추가
+ 인벤토리 비어있을때 딕셔너리 오류나던거 수정
+ 데미지 float형으로 수정해서 오류 수정
+ InventoryManager 전역변수로 선언해서 불필요한 사용 줄임

+++ SpartaDungeon/Skill.cs
@@ -16,7 +16,7 @@ namespace SpartaDungeon
 
         Random random = new Random();
 
-        public Skill(string name, string desc, int mp, int damage = 0)
+        public Skill(string name, string desc, int mp, float damage = 0.0f)
         {
             Name = name;
             Desc = desc;
@@ -124,7 +124,7 @@ namespace SpartaDungeon
         {
 
             Console.WriteLine($"{Name}를 뽑았습니다");
-            Console.WriteLine($"\"파랑이 좋겠어\"");            
+            Console.WriteLine($"\"파랑이 좋겠어\"");
 
         }
